Moving speedup unit tests into `test__wrap*` files.
