+++ tests/unit/test__py_curve_helpers.py
@@ -21,6 +21,7 @@ except ImportError:  # pragma: NO COVER
     SCIPY_INT = None
 from tests.unit import utils
 
+
 FLOAT64 = np.float64  # pylint: disable=no-member
 SPACING = np.spacing  # pylint: disable=no-member
 
@@ -166,15 +167,6 @@ class Test__subdivide_nodes(utils.NumPyTestCase):
         self._points_check(nodes)
 
 
-@utils.needs_speedup
-class Test_speedup_subdivide_nodes(Test__subdivide_nodes):
-    @staticmethod
-    def _call_function_under_test(nodes):
-        from bezier import _speedup
-
-        return _speedup.subdivide_nodes_curve(nodes)
-
-
 class Test__evaluate_multi_barycentric(utils.NumPyTestCase):
     @staticmethod
     def _call_function_under_test(nodes, lambda1, lambda2):
@@ -199,17 +191,6 @@ class Test__evaluate_multi_barycentric(utils.NumPyTestCase):
             ]
         )
         self.assertEqual(result, expected)
-
-
-@utils.needs_speedup
-class Test_speedup_evaluate_multi_barycentric(
-    Test__evaluate_multi_barycentric
-):
-    @staticmethod
-    def _call_function_under_test(nodes, lambda1, lambda2):
-        from bezier import _speedup
-
-        return _speedup.evaluate_multi_barycentric(nodes, lambda1, lambda2)
 
 
 class Test__evaluate_multi(utils.NumPyTestCase):
@@ -356,15 +337,6 @@ class Test__evaluate_multi(utils.NumPyTestCase):
         self.assertEqual(expected, binomial_coefficients)
 
 
-@utils.needs_speedup
-class Test_speedup_evaluate_multi(Test__evaluate_multi):
-    @staticmethod
-    def _call_function_under_test(nodes, s_vals):
-        from bezier import _speedup
-
-        return _speedup.evaluate_multi(nodes, s_vals)
-
-
 class Test_vec_size(unittest.TestCase):
     @staticmethod
     def _call_function_under_test(nodes, s_val):
@@ -444,25 +416,6 @@ class Test__compute_length(unittest.TestCase):
                 self._call_function_under_test(nodes)
 
 
-@utils.needs_speedup
-class Test_speedup_compute_length(Test__compute_length):
-    @staticmethod
-    def _call_function_under_test(nodes):
-        from bezier import _speedup
-
-        return _speedup.compute_length(nodes)
-
-    def _scipy_skip(self):
-        # Fortran implementation directly includes QUADPACK, so the presence
-        # or absence of SciPy is irrelevant.
-        pass
-
-    def test_without_scipy(self):
-        # Fortran implementation directly includes QUADPACK, so the presence
-        # or absence of SciPy is irrelevant.
-        pass
-
-
 class Test__elevate_nodes(utils.NumPyTestCase):
     @staticmethod
     def _call_function_under_test(nodes):
@@ -489,15 +442,6 @@ class Test__elevate_nodes(utils.NumPyTestCase):
             ]
         )
         self.assertEqual(result, expected)
-
-
-@utils.needs_speedup
-class Test_speedup_elevate_nodes(Test__elevate_nodes):
-    @staticmethod
-    def _call_function_under_test(nodes):
-        from bezier import _speedup
-
-        return _speedup.elevate_nodes(nodes)
 
 
 class Test_de_casteljau_one_round(utils.NumPyTestCase):
@@ -566,15 +510,6 @@ class Test__specialize_curve(utils.NumPyTestCase):
         self.assertEqual(result, expected)
 
 
-@utils.needs_speedup
-class Test_speedup_specialize_curve(Test__specialize_curve):
-    @staticmethod
-    def _call_function_under_test(nodes, start, end):
-        from bezier import _speedup
-
-        return _speedup.specialize_curve(nodes, start, end)
-
-
 class Test__evaluate_hodograph(utils.NumPyTestCase):
     @staticmethod
     def _call_function_under_test(s, nodes):
@@ -619,15 +554,6 @@ class Test__evaluate_hodograph(utils.NumPyTestCase):
             self.assertEqual(first_deriv[1, 0], y_prime)
 
 
-@utils.needs_speedup
-class Test_speedup_evaluate_hodograph(Test__evaluate_hodograph):
-    @staticmethod
-    def _call_function_under_test(s, nodes):
-        from bezier import _speedup
-
-        return _speedup.evaluate_hodograph(s, nodes)
-
-
 class Test__get_curvature(unittest.TestCase):
     @staticmethod
     def _call_function_under_test(nodes, tangent_vec, s):
@@ -663,15 +589,6 @@ class Test__get_curvature(unittest.TestCase):
         self.assertEqual(result, -4.0)
 
 
-@utils.needs_speedup
-class Test_speedup_get_curvature(Test__get_curvature):
-    @staticmethod
-    def _call_function_under_test(nodes, tangent_vec, s):
-        from bezier import _speedup
-
-        return _speedup.get_curvature(nodes, tangent_vec, s)
-
-
 class Test__newton_refine(unittest.TestCase):
     @staticmethod
     def _call_function_under_test(nodes, point, s):
@@ -687,15 +604,6 @@ class Test__newton_refine(unittest.TestCase):
         point = np.asfortranarray([[1.75], [0.625], [1.625]])
         new_s = self._call_function_under_test(nodes, point, 0.25)
         self.assertEqual(110.0 * new_s, 57.0)
-
-
-@utils.needs_speedup
-class Test_speedup_newton_refine(Test__newton_refine):
-    @staticmethod
-    def _call_function_under_test(nodes, point, s):
-        from bezier import _speedup
-
-        return _speedup.newton_refine_curve(nodes, point, s)
 
 
 class Test__locate_point(unittest.TestCase):
@@ -740,15 +648,6 @@ class Test__locate_point(unittest.TestCase):
         point = np.asfortranarray([[2.0], [0.0]])
         result = self._call_function_under_test(nodes, point)
         self.assertEqual(result, 1.0)
-
-
-@utils.needs_speedup
-class Test_speedup_locate_point(Test__locate_point):
-    @staticmethod
-    def _call_function_under_test(nodes, point):
-        from bezier import _speedup
-
-        return _speedup.locate_point_curve(nodes, point)
 
 
 class Test__reduce_pseudo_inverse(utils.NumPyTestCase):
@@ -838,15 +737,6 @@ class Test__reduce_pseudo_inverse(utils.NumPyTestCase):
             self._call_function_under_test(nodes)
         self.assertEqual(exc_info.exception.degree, degree)
         self.assertEqual(exc_info.exception.supported, (1, 2, 3, 4))
-
-
-@utils.needs_speedup
-class Test_speedup__reduce_pseudo_inverse(Test__reduce_pseudo_inverse):
-    @staticmethod
-    def _call_function_under_test(nodes):
-        from bezier import _speedup
-
-        return _speedup.reduce_pseudo_inverse(nodes)
 
 
 class Test_projection_error(unittest.TestCase):
@@ -984,12 +874,3 @@ class Test__full_reduce(utils.NumPyTestCase):
             self._call_function_under_test(nodes)
         self.assertEqual(exc_info.exception.degree, degree)
         self.assertEqual(exc_info.exception.supported, (0, 1, 2, 3, 4))
-
-
-@utils.needs_speedup
-class Test_speedup_full_reduce(Test__full_reduce):
-    @staticmethod
-    def _call_function_under_test(nodes):
-        from bezier import _speedup
-
-        return _speedup.full_reduce(nodes)
