+++ tests/unit/test__py_intersection_helpers.py
@@ -722,15 +722,6 @@ class Test_full_newton(unittest.TestCase):
         utils.almost(self, 0.5, computed_t, 1)
 
 
-@utils.needs_speedup
-class Test_speedup_newton_refine(Test__newton_refine):
-    @staticmethod
-    def _call_function_under_test(s, nodes1, t, nodes2):
-        from bezier import _speedup
-
-        return _speedup.newton_refine_curve_intersect(s, nodes1, t, nodes2)
-
-
 class TestIntersection(unittest.TestCase):
     @staticmethod
     def _get_target_class():
